requirements.txt
google-genai
chromadb
python-dotenv
fastapi
uvicorn
pydantic
pyyaml